+++ CHANGELOG.md
@@ -9,7 +9,8 @@
 
 ## Tagged Releases
 
-- [unreleased](https://github.com/Seika139/discord-notify/compare/v0.2.0...HEAD)
+- [unreleased](https://github.com/Seika139/discord-notify/compare/v0.3.0...HEAD)
+- [0.3.0](https://github.com/Seika139/discord-notify/compare/v0.2.0...v0.3.0)
 - [0.2.0](https://github.com/Seika139/discord-notify/compare/v0.1.5...v0.2.0)
 - [0.1.5](https://github.com/Seika139/discord-notify/compare/v0.1.4...v0.1.5)
 - [0.1.4](https://github.com/Seika139/discord-notify/compare/v0.1.3...v0.1.4)
@@ -19,6 +20,8 @@
 - [0.1.0](https://github.com/Seika139/discord-notify/releases/tag/v0.1.0)
 
 ## [Unreleased]
+
+## [0.3.0] - 2026-06-29
 
 ### Added